Updating the WLAN Calibration Script
This make sure that it will only calibrate on R5 based driver.
Even if we hack the module to be loaded to wlcore_sdio.ko to use WiLink8
in init.rc and user has WiLink7, we can still use the script fo calibration

If the calibartion fails it will restore the previous nvs file.

Change-Id: I6e5a100ef7485ec5d408ad33f1cd430b2128c995
Signed-off-by: Panduranga Mallireddy <panduranga_mallireddy@ti.com>